Fix bug where initial description is seeded with unrelated commits.
If there are commits on the upstream branches which are after the
merge-base, they will get included in the description.

The bug is related to the fact that master... has different meaning
in git-log and git-diff. In git-diff, master... will show the diff
between HEAD and the merge-base. For git-log, it will show commits
that are in master or HEAD but not in both. These can be different.